requirements.txt
asgiref==3.8.1
Django==5.0.6
django-custom-user==1.1
django-debug-toolbar==4.4.2
django-web-components==0.2.0
djangorestframework==3.15.2
docutils==0.21.2
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
